Get the responseType and supportType from the collection.
Previously, we were injecting the support/response type by overriding 'add' on the PlaceCollection, however if we called 'create' to add the place, the responseType and supportType would never get set on the model.  So, instead of overriding add to pass the types in as options, just reach back to the place's collection if it doesn't find the types as options passed in directly.