Upgrade Mapbox GL JS to 1.4.1 to support satellite layers
Currently using a layer with `type: mapbox` and a tile-based styles such as
satellite imagery fails with an error like:

```
Unknown expression "image". If you wanted a literal array, use ["literal", [...]].
````

Support for the `image` expression was added to Mapbox GL JS in version 1.4, so
I've conservatively upgraded from 0.53.0 to the latest 1.4 patch release.

While using a satellite layer with the normal `url:` config is still possible,
we've found that Mapbox's satellite styles can't be zoomed past 17 or 18 and are
billed at a higher rate.