Add django-redis dependency and upgrade to support Django 3.2
Given that Redis config appears in the default settings, it seems appropriate to
include it as a dependency.

The default Redis config in `settings.py` also isn't compatible with a
django-redis that supports Django 3.2, so I've simplified the config. My
thinking is that the defaults should be simple and more complex would be best
done case-by-case in `local_settings.py`.