Make radio and checkbox group labels relatively positioned
This fixes the phenomenon of the screen jumping up when you choose a check box low down in the list. The problem was that the actual input element would be positioned near the top of the group because it's absolutely positioned (in order to be removed from the normal document flow). To solve the issue, we make the label itself relatively positioned, ensuring that the input element is positioned near the label.